Ranges support

+++ src/test/php/web/unittest/handler/FilesFromTest.class.php
@@ -42,7 +42,7 @@ class FilesFromTest extends \unittest\TestCase {
   /** @return void */
   public function tearDown() {
     foreach ($this->cleanup as $folder) {
-      $folder->unlink();
+      $folder->exists() && $folder->unlink();
     }
   }
 
@@ -61,6 +61,7 @@ class FilesFromTest extends \unittest\TestCase {
 
     $this->assertResponse(
       "HTTP/1.1 200 OK\r\n".
+      "Accept-Ranges: bytes\r\n".
       "Last-Modified: <Date>\r\n".
       "Content-Type: text/html\r\n".
       "Content-Length: 4\r\n".
@@ -95,6 +96,7 @@ class FilesFromTest extends \unittest\TestCase {
 
     $this->assertResponse(
       "HTTP/1.1 200 OK\r\n".
+      "Accept-Ranges: bytes\r\n".
       "Last-Modified: <Date>\r\n".
       "Content-Type: text/html\r\n".
       "Content-Length: 4\r\n".
@@ -136,6 +138,129 @@ class FilesFromTest extends \unittest\TestCase {
       "Content-Length: 26\r\n".
       "\r\n".
       "The file '/' was not found",
+      $out->bytes
+    );
+  }
+
+  #[@test, @values([
+  #  ['0-3', 'Home'],
+  #  ['4-7', 'page'],
+  #  ['0-0', 'H'],
+  #  ['4-4', 'p'],
+  #  ['7-7', 'e']
+  #])]
+  public function range_with_start_and_end($range, $result) {
+    $in= new TestInput('GET', '/', ['Range' => 'bytes='.$range]);
+    $out= new TestOutput(); 
+
+    $files= (new FilesFrom($this->pathWith(['index.html' => 'Homepage'])));
+    $files->handle(new Request($in), new Response($out));
+
+    $this->assertResponse(
+      "HTTP/1.1 206 Partial Content\r\n".
+      "Accept-Ranges: bytes\r\n".
+      "Last-Modified: <Date>\r\n".
+      "Content-Type: text/html\r\n".
+      "Content-Range: bytes ".$range."/8\r\n".
+      "Content-Length: ".strlen($result)."\r\n".
+      "\r\n".
+      $result,
+      $out->bytes
+    );
+  }
+
+  #[@test]
+  public function range_from_offset_until_end() {
+    $in= new TestInput('GET', '/', ['Range' => 'bytes=4-']);
+    $out= new TestOutput(); 
+
+    $files= (new FilesFrom($this->pathWith(['index.html' => 'Homepage'])));
+    $files->handle(new Request($in), new Response($out));
+
+    $this->assertResponse(
+      "HTTP/1.1 206 Partial Content\r\n".
+      "Accept-Ranges: bytes\r\n".
+      "Last-Modified: <Date>\r\n".
+      "Content-Type: text/html\r\n".
+      "Content-Range: bytes 4-7/8\r\n".
+      "Content-Length: 4\r\n".
+      "\r\n".
+      "page",
+      $out->bytes
+    );
+  }
+
+  #[@test, @values([0, 8192, 10000])]
+  public function range_last_four_bytes($offset) {
+    $in= new TestInput('GET', '/', ['Range' => 'bytes=-4']);
+    $out= new TestOutput(); 
+
+    $files= (new FilesFrom($this->pathWith(['index.html' => str_repeat('*', $offset).'Homepage'])));
+    $files->handle(new Request($in), new Response($out));
+
+    $this->assertResponse(
+      "HTTP/1.1 206 Partial Content\r\n".
+      "Accept-Ranges: bytes\r\n".
+      "Last-Modified: <Date>\r\n".
+      "Content-Type: text/html\r\n".
+      "Content-Range: bytes ".($offset + 4)."-".($offset + 7)."/".($offset + 8)."\r\n".
+      "Content-Length: 4\r\n".
+      "\r\n".
+      "page",
+      $out->bytes
+    );
+  }
+
+  #[@test, @values([
+  #  'bytes=0-2000',
+  #  'bytes=4-2000',
+  #  'bytes=2000-',
+  #  'bytes=2000-2001',
+  #  'bytes=2000-0',
+  #  'bytes=4-0',
+  #  'characters=0-'
+  #])]
+  public function range_unsatisfiable($range) {
+    $in= new TestInput('GET', '/', ['Range' => $range]);
+    $out= new TestOutput();
+
+    $files= (new FilesFrom($this->pathWith(['index.html' => 'Homepage'])));
+    $files->handle(new Request($in), new Response($out));
+
+    $this->assertResponse(
+      "HTTP/1.1 416 Range Not Satisfiable\r\n".
+      "Accept-Ranges: bytes\r\n".
+      "Last-Modified: <Date>\r\n".
+      "Content-Range: bytes */8\r\n".
+      "\r\n",
+      $out->bytes
+    );
+  }
+
+  #[@test]
+  public function multi_range() {
+    $in= new TestInput('GET', '/', ['Range' => 'bytes=0-3,4-7']);
+    $out= new TestOutput(); 
+
+    $files= (new FilesFrom($this->pathWith(['index.html' => 'Homepage'])));
+    $files->handle(new Request($in), new Response($out));
+
+    $this->assertResponse(
+      "HTTP/1.1 206 Partial Content\r\n".
+      "Accept-Ranges: bytes\r\n".
+      "Last-Modified: <Date>\r\n".
+      "Content-Type: multipart/byteranges; boundary=594fa07300f865fe\r\n".
+      "Content-Length: 186\r\n".
+      "\r\n".
+      "\r\n--594fa07300f865fe\r\n".
+      "Content-Type: text/html\r\n".
+      "Content-Range: bytes 0-3/8\r\n\r\n".
+      "Home".
+      "\r\n--594fa07300f865fe\r\n".
+      "Content-Type: text/html\r\n".
+      "Content-Range: bytes 4-7/8\r\n\r\n".
+      "page".
+      "\r\n--594fa07300f865fe--\r\n",
       $out->bytes
     );
   }
